Added in a new movement behavior based on Craig Reynold's Steering Behavior.
The first one is Seek, the introductory behavior of steering AI. Requires the two velocities: Desired velocity - based on direction from object to target, and Steering velocity - the difference of object's velocity and desired velocity. For the smooth turning to work on the seek steering, a scalable constant is applied when multiplying the steering velocity before adding it to the overall velocity of the object.